SUMMARY:   Fixed bug in sort used by Zoltan_Verify_Graph
     BUGS:      NA

     CHECKLIST:

     1)   Did you add any source files?  no

     2)   Did you update or change the build system?  no

     3) Did you rerun the test suite on at least the packages that you
        changed?   yes

     COMMENTS:

In the sort used by Zoltan_Verify_Graph,
the reordering of the auxiliary array was incorrect in the
special case where the primary array had duplicate entries.
For Zoltan_Verify_Graph, this case would be very rare, as it
would indicate that a vertex A had more than one edge to
another vertex B (i.e., B would be listed twice in the
primary array).

(checklist version 2007-06-20)