SUMMARY: Adding failing test for new Rythmos adjoint solver
BUGS: 3889

CHECKLIST:

1) Did you add any source or other files needed for building, testing,
   or documentation? Yes
1.a) Did you do a 'cvs add'? Yes
1.b) Did you add them to a Makefile.am file in some way (either as
     a header or soruce file or to EXTRA_DIST)? Yes
1.c) Did you do a new bootstrap? Yes

2) Did you update or change the build system (other than just adding
   source files)? Yes
2.a) Did you remember to do a bootstrap and do a 'cvs add' on any
     Makefile.am files and Makefile.in files you created? Yes

3) Did you do a cvs update *very* recently before rebuilding and
   retesting? Yes

4) If justified, did you type 'make partial-clean' and then rebuild all
   libraries and executables with 'make everything'? No, I only rebuilt
   rythmos

5) Did you rerun the test suite on at least the packages that you
   changed (rerunning the entire Trlinos test suite on all affected
   packages is *strongly* recommended)? Yes, but I only rebuilt rytymos
   and ran its tests.

COMMENTS:

Here I am adding just the skeleton for a new adjoint capability for Rythmos.
I am checking in a failing test that has a test condition that checks for a
specific type of failure as passing.

I am also checking in an uncompiled skeleton for a LOWSB decorator class for
an adjoint linear operator solver.  I will write a proper unit test for this
and fill this out next.