SUMMARY: Move config headers for SIERRA/BJAM and the list of headers to link into Trilinos.
BUGS: 3954

CHECKLIST:

1) Did you add any source or other files needed for building, testing,
   or documentation? No, not for the native Trilinos build system

2) Did you update or change the build system (other than just adding
   source files)? No, not for the native Trilnos build system

3) Did you do a cvs update *very* recently before rebuilding and
   retesting? Yes, but not needed

4) If justified, did you type 'make partial-clean' and then rebuild all
   libraries and executables with 'make everything'? No, this does not
   affect the native Trilinos build sytem

5) Did you rerun the test suite on at least the packages that you
   changed (rerunning the entire Trlinos test suite on all affected
   packages is *strongly* recommended)? No, this does not affect
   the native Trilinos build system

COMMENTS:

Here I am moving the files trilinos_link_headers.txt and config_headers/*.h
from Nbtools/Trilinos/[dev,9.0] into the main Trilinos CVS reposotry.  The
reason that I am doing this is to avoid having to maintain two copies these
files before the Trilinos 9.0 branch and to clean out the Nbtools repository
as requested by the SIERRA tools devleopers.

These are not a lot of files and is therefore not a lot of data to store in
the Trilinos CVS repository.