SUMMARY:   removed 'unused parameter' warnings, incorrect parameter types
     CHECKLIST:

     1)   Did you add any source files?  no

     2)   Did you update or change the build system?  no

     3) Did you rerun the test suite on at least the packages that you
        changed?  yes

     COMMENTS:

In DefaultMpiComm and MpiReducitonOpSetter, added some '(void)param;' statements to
eliminate '--ansi' 'unused parameter' warnings

In CommHelpers, some immediates ('1' and '0') were interpretted by my compiler as ints.
When the helper routine was templated on non-int Ordinal, this prevented the compiler from
implicitly determining the routine to call. I explicitly stated the template arguments to
the called helper functions; the immediate are implicitly cast to the Ordinal type.

(checklist version 2007-06-20)