Bug 3990.
This removes almost all dependencies on the nox blas/lapack wrappers in favor of the teuchos blas/lapack wrappers.  There is one final dependency in a loca file on a lapack method that the teuchos wrappers do not currently support. One this is added to Teuchos, we can remove the nox wrappers.