SUMMARY: Setting errorTolSlack > 1.0 instead of < 1.0 :-)
BUGS: 4000

CHECKLIST:

1) Did you add any source or other files needed for building, testing,
   or documentation? No

2) Did you update or change the build system (other than just adding
   source files)? No

3) Did you do a cvs update *very* recently before rebuilding and
   retesting? Yes

4) If justified, did you type 'make partial-clean' and then rebuild all
   libraries and executables with 'make everything'? No, I just rebuilt
   the tests suites for teuchos and rtop for serial debug and mpi
   optimized

5) Did you rerun the test suite on at least the packages that you
   changed (rerunning the entire Trlinos test suite on all affected
   packages is *strongly* recommended)? Yes, but only for teuchos
   and rtop for serial debug and mpi optimized (all passed)

COMMENTS:

I accedentally set errorTolSlack = 1e-2.  This number is multiplied by machine
epsilon which means that if you have any round-off that the tests will fail!
On my linux machine where I tested this, there was no round-off and all the
errors where zero so the test passed.

However, on s858353, I got:

47. ROpNorm2_float_unitStride_UnitTest ...
 v=-0.878871

 Check: rel_err(norm2Op(*norm2), SMT::squareroot(n)*ST::magnitude(v))
        = rel_err(1.75774, 1.75774) = 6.78195e-08
          <= tol = 1.19209e-09 : failed
 [FAILED]
 Location: ../../../../../packages/rtop/test/ops_lib/ROpNorm2_UnitTests.cpp:9

And on beowulf I got:

3. ROpDotProd_ComplexFloat_nonunitStride_UnitTest ...
 v1=(-0.270431,0.0268018), v2=(0.904459,0.83239)

 Check: rel_err(dotProdOp(*dotProd), ST::conjugate(v1)*v2*as<Scalar>(n))
        = rel_err((-0.889137,-0.997381), (-0.889137,-0.997381)) = 4.46088e-08
          <= tol = 1.19209e-09 : failed
 [FAILED]
 Location:
../../../../../packages/rtop/test/ops_lib/ROpDotProd_UnitTests.cpp:33

11. ROpDotProd_float_unitStride_UnitTest ...
 v1=0.275105, v2=0.0485744

 Check: rel_err(dotProdOp(*dotProd), ST::conjugate(v1)*v2*as<Scalar>(n))
        = rel_err(0.0534521, 0.0534521) = 6.96938e-08
          <= tol = 1.19209e-09 : failed
 [FAILED]
 Location: ../../../../../packages/rtop/test/ops_lib/ROpDotProd_UnitTests.cpp:9

What I ment to do was to set errorTolSlack > 1 to *loosen* the tolerance above
machine epsilon.

Therefore, I have set errorTolSolack = 1e+1 to allow some accumulation of
error.

I also removed the file opsUnitTests.cpp since it got broken up and is not
being used anymore.  It was not being included in Makefile.am so there was
nothing else to do but to remove this.

This should fix all of the RTOp tests on all platforms ...