Added the line:
 solvePrecOpPtr = Teuchos::null;

at the end of destroyPreconditioner() to force the preconditioner to actually
be destroyed.  Without this, destroyPreconditioner() wouldn't actually destroy
the preconditioner and it wouldn't be destroyed until the next one was created
by createPreconditioner().  There isn't anything wrong with this per se, but
it is misleading and can make tracking errors down more difficult.

All tests pass.